+++ src/test/kotlin/com/frejdh/util/environment/KotlinSupportTests.kt
@@ -10,7 +10,7 @@ class KotlinSupportTests {
         Assert.assertEquals(50, Config.getInteger("env.test1.test-of-env-int1", 10)) // Should be found
         Assert.assertEquals(50, Config.getInteger("property.does.not.exist", 50)) // Correct default value
         Assert.assertNull(Config.getInteger("property.does.not.exist"))
-        Assert.assertThrows<IllegalArgumentException>(IllegalArgumentException::class.java) { Config.get("env.test1.test-of-env-int1", PropertyTests::class.java) } // Should throw
+        Assert.assertThrows<IllegalArgumentException>(IllegalArgumentException::class.java) { Config.get("env.test1.test-of-env-int1", PropertyParserTests::class.java) } // Should throw
     }
 }